requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
google-genai
pinecone
python-jose[cryptography]
passlib[bcrypt]

ffmpeg-python
